New methods for the Commander, and new operator: @.
Commander now has the following methods:
- tap and untap, with both calling the respective methods for the master
  clock.
- value, which sets a variable's value.

The write method also works in tandem with the new operator: it can
receive an additional parameter 'm' such that the grid origin for the
text block to be written will be the position of the marker 'm' defined
by the @ operator. That is: @ creates a named marker on the grid to be
used as an origin for 'write'.